Split downloader from file_utils.py.
